Question to creater
Good day!
I apologize for the machine translation and for such an inappropriate attempt to contact you.
As a fan of the game with experience, I have a sharp question: how to transfer the character from the "campaign" to the "adventure mod"?
Standard methods did not work = (
Perhaps you had ideas on this subject? Maybe there is a ready-made solution?
Thanks in advance for your reply. Just in case, my E-mail: bel.fantom@gmail.com